requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.6
redis>=4.5.4
celery>=5.2.7
django-celery-beat>=2.5.0
django-redis>=5.2.0
Pillow>=9.5.0
python-dotenv>=1.0.0
gunicorn>=20.1.0
stripe>=5.4.0 